% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_boundary.R
\name{detect_boundary_num}
\alias{detect_boundary_num}
\title{Detect the range for numerical variables}
\usage{
detect_boundary_num(x, ...)
}
\arguments{
\item{x}{\code{\link{validator}} object, rule set to be checked}

\item{...}{currently not used}
}
\value{
\code{\link{data.frame}} with columns "variable", "lowerbound", "upperbound".
}
\description{
Detect for each numerical variable in a validation rule set, what its maximum and minimum values are.
This allows for manual rule set checking: does rule set \code{x} overly constrain numerical values?
}
\details{
This procedure only finds minimum and maximum values, but misses gaps.
}
\examples{
rules <- validator(
  x >= 1,
  x + y <= 10,
  y >= 6
)

detect_boundary_num(rules)
}
\references{
Statistical Data Cleaning with R (2017), Chapter 8, M. van der Loo, E. de Jonge

Simplifying constraints in data editing (2015). Technical Report 2015|18, Statistics Netherlands, J. Daalmans
}
\seealso{
\code{\link{detect_fixed_values}}
}
