% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_fixed_values.R
\name{detect_fixed_values}
\alias{detect_fixed_values}
\title{Detect fixed values}
\usage{
detect_fixed_values(x, ...)
}
\arguments{
\item{x}{\code{\link{validator}} object with the validation rules.}

\item{...}{not used.}
}
\description{
Detect fixed values
}
\examples{
\dontrun{
library(validate)
rules <- validator( x >= 0
                  , x <= 0
                  )
detect_fixed_values(rules)
simplify_fixed_values(rules)
}

  # rules <- validate::validator( x1 + x2 + x3 == 0
  #                               , x1 + x2 >= 0
  #                               , x3 >= 0
  # )
}
\seealso{
\code{\link{simplify_fixed_values}}
}
