% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_infeasible_if_rules.R
\name{detect_infeasible_if_rules}
\alias{detect_infeasible_if_rules}
\title{Detect infeasible if clauses}
\usage{
detect_infeasible_if_rules(x, ..., verbose = TRUE)
}
\arguments{
\item{x}{A validator object.}

\item{...}{Additional arguments passed to \code{detect_if_clauses}.}

\item{verbose}{Logical. If \code{TRUE}, print the results.}
}
\value{
A list of contradictions found in the if clauses, or \code{NULL} if none are found.
}
\description{
Detect if clauses that contradict each other. This is useful to detect if clauses that are not satistifable.
}
\examples{
rules <- validator(
  if (nace == "a") export == "y",
  if (nace == "a")  export == "n"
)

detect_infeasible_if_rules(rules)
}
\seealso{
Other feasibility: 
\code{\link{detect_boundary_cat}()},
\code{\link{detect_boundary_num}()},
\code{\link{detect_infeasible_rules}()},
\code{\link{is_contradicted_by}()},
\code{\link{is_infeasible}()},
\code{\link{make_feasible}()}
}
\concept{feasibility}
