% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_conditional.R
\name{simplify_conditional}
\alias{simplify_conditional}
\title{Simplify conditional statements}
\usage{
simplify_conditional(x, ...)
}
\description{
Conditional rules may be constrained by the others rules in a validation rule set.
This procedure tries to simplify conditional statements.
}
\examples{
library(validate)

# non-relaxing clause
rules <- validator( r1 = if (x > 1) y > 3
                  , r2 = y < 2
                  )
# y > 3 is always FALSE so r1 can be simplified
simplify_conditional(rules)


# non-constraining clause
rules <- validator( r1 = if (x > 0) y > 0
                  , r2 = if (x < 1) y > 1
                  )
simplify_conditional(rules)
}
\references{
TODO non-constraining, non-relaxing
}
