% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_fixed_values.R
\name{simplify_fixed_values}
\alias{simplify_fixed_values}
\title{Simplify fixed values}
\usage{
simplify_fixed_values(x, ...)
}
\arguments{
\item{x}{\code{\link{validator}} object with validation rules}

\item{...}{passed to \code{\link{substitute_values}}.}
}
\value{
\code{\link{validator}} object in which
}
\description{
Simplify fixed values
}
