% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_boundary.R
\name{detect_boundary_cat}
\alias{detect_boundary_cat}
\title{Detect viable domains for categorical variables}
\usage{
detect_boundary_cat(x, ...)
}
\arguments{
\item{x}{\code{\link{validator}} object with rules}

\item{...}{not used}
}
\description{
Detect viable domains for categorical variables
}
