% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_fixed_values.R
\name{detect_fixed_values}
\alias{detect_fixed_values}
\title{Detect fixed values}
\usage{
detect_fixed_values(x, eps = x$options("lin.eq.eps"), ...)
}
\arguments{
\item{x}{\code{\link{validator}} object with the validation rules.}

\item{eps}{detected fixed values will have this precission.}

\item{...}{not used.}
}
\description{
Detects fixed values in the rule set. To simplify a rule set, these fixed values can be substituted.
}
\examples{
library(validate)
rules <- validator( x >= 0
                  , x <= 0
                  )
detect_fixed_values(rules)
simplify_fixed_values(rules)

rules <- validator( x1 + x2 + x3 == 0
                  , x1 + x2 >= 0
                  , x3 >= 0
                  )
simplify_fixed_values(rules)
}
\seealso{
\code{\link{simplify_fixed_values}}

Other redundancy: \code{\link{detect_redundancy}},
  \code{\link{is_implied_by}},
  \code{\link{simplify_fixed_values}},
  \code{\link{simplify_redundancy}},
  \code{\link{simplify_rules}}
}
