% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasible.R
\name{detect_infeasible_rules}
\alias{detect_infeasible_rules}
\title{Detect which rules cause infeasibility}
\usage{
detect_infeasible_rules(x, weight = numeric(), ...)
}
\arguments{
\item{x}{\code{\link{validator}} object with rules}

\item{weight}{optional named \code{\link{numeric}} with weights. The names}

\item{...}{not used}
}
\value{
\code{character} with the names of the rules that are causing infeasibility.
}
\description{
Detect which rules cause infeasibility. This methods tries to remove the minimum number of rules to make the system
mathematically feasible. Note that this may not give you the desired system, because some rules may be more important
to you than others. This can be mitigated by supplying weights for the rules.
}
\examples{
rules <- validator( x > 0)

is_infeasible(rules)

rules <- validator( rule1 = x > 0
                  , rule2 = x < 0
                  )

is_infeasible(rules)

detect_infeasible_rules(rules)
make_feasible(rules)

# find out the conflict with this rule
is_contradicted_by(rules, "rule1")
}
