% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redundancy.R
\name{detect_redundancy}
\alias{detect_redundancy}
\title{Detect redundant rules without removing.}
\usage{
detect_redundancy(x, ...)
}
\description{
Detect redundant rules without removing.
}
\note{
For removal of duplicate rules, simplify
}
\examples{
rules <- validator( rule1 = x > 1
                  , rule2 = x > 2
                  )

# rule1 is superfluous
simplify_redundancy(rules)

rules <- validator( rule1 = x > 2
                  , rule2 = x > 2
)

# standout: rule1 and rule2, oldest rules wins
simplify_redundancy(rules)

# Note that detection signifies both rules!
detect_redundancy(rules)
}
