% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contradicted.R
\name{is_contradicted_by}
\alias{is_contradicted_by}
\title{Find out for a contradicting}
\usage{
is_contradicted_by(x, rule_name)
}
\arguments{
\item{x}{\code{\link{validator}} object with rules.}

\item{rule_name}{\code{character} with the names of the rules that are causing infeasibility.}
}
\description{
Find out for a contradicting
}
