% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasible.R
\name{is_infeasible}
\alias{is_infeasible}
\title{Check the feasibility of a rule set}
\usage{
is_infeasible(x, ...)
}
\arguments{
\item{x}{\code{validator} object with validation rules.}

\item{...}{not used}
}
\value{
TRUE or FALSE
}
\description{
This function checks if the linear, categorical and conditional rules are consistent. If not
the system is said to be "infeasible" and it is not possible to validate any data with it, because
there will be always a rule that is not obeyed.
}
\details{
It works by translating these rules into a mixed integer problem and check this system of
(in)equalities for feasibility.
}
