% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasible.R
\name{make_feasible}
\alias{make_feasible}
\title{Make an infeasible system feasible.}
\usage{
make_feasible(x, ...)
}
\arguments{
\item{x}{\code{\link{validator}} object with the validation rules.}

\item{...}{passed to \code{\link{detect_infeasible_rules}}}
}
\value{
\code{\link{validator}} object with feasible rules.
}
\description{
Make an infeasible system feasible.
}
\examples{
rules <- validator( x > 0)

is_infeasible(rules)

rules <- validator( rule1 = x > 0
                  , rule2 = x < 0
                  )

is_infeasible(rules)

detect_infeasible_rules(rules)
make_feasible(rules)

# find out the conflict with this rule
is_contradicted_by(rules, "rule1")
}
