% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasible.R
\name{make_feasible}
\alias{make_feasible}
\title{If a system is infeasible}
\usage{
make_feasible(x, ...)
}
\arguments{
\item{x}{\code{\link{validator}} object with the validation rules.}

\item{...}{not used.}

\item{x}{\code{validator} object with validation rules.}
}
\description{
If a system is infeasible
}
