% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasible.R
\name{make_feasible}
\alias{make_feasible}
\title{Make an infeasible system feasible.}
\usage{
make_feasible(x, ...)
}
\arguments{
\item{x}{\code{\link{validator}} object with the validation rules.}

\item{...}{not used.}
}
\description{
Make an infeasible system feasible.
}
