% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_clauses.R
\name{simplify_conditional}
\alias{simplify_conditional}
\title{Simplify conditional statements}
\usage{
simplify_conditional(x)
}
\description{
Conditional rules may be constrained by the others rules in a validation rule set.
This procedure tries to
}
