% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_fixed_values.R
\name{simplify_fixed_values}
\alias{simplify_fixed_values}
\title{Simplify fixed variables}
\usage{
simplify_fixed_values(x, eps = 1e-08, ...)
}
\arguments{
\item{x}{\code{\link{validator}} object with validation rules}

\item{eps}{detected fixed values will have this precission.}

\item{...}{passed to \code{\link{substitute_values}}.}
}
\value{
\code{\link{validator}} object in which
}
\description{
Detect variables of which the values are restricted to a single value by the
rule set. Simplify the rule set by replacing fixed variables with these values.
}
\examples{
\dontrun{
library(validate)
rules <- validator( x >= 0
                  , x <= 0
                  )
detect_fixed_values(rules)
simplify_fixed_values(rules)

rules <- validator( x1 + x2 + x3 == 0
                  , x1 + x2 >= 0
                  , x3 >= 0
                  )
simplify_fixed_values(rules)
}
}
\seealso{
Other redundancy: \code{\link{detect_fixed_values}},
  \code{\link{detect_redundancy}},
  \code{\link{is_implied_by}},
  \code{\link{simplify_redundancy}},
  \code{\link{simplify_rules}}
}
