% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_fixed_values.R
\name{simplify_fixed_values}
\alias{simplify_fixed_values}
\title{Simplify fixed values}
\usage{
simplify_fixed_values(x, eps = 1e-08, ...)
}
\arguments{
\item{x}{\code{\link{validator}} object with validation rules}

\item{eps}{detected fixed values will have this precission.}

\item{...}{passed to \code{\link{substitute_values}}.}
}
\value{
\code{\link{validator}} object in which
}
\description{
Simplify fixed values
}
\examples{
\dontrun{
library(validate)
rules <- validator( x >= 0
                  , x <= 0
                  )
detect_fixed_values(rules)
simplify_fixed_values(rules)

rules <- validator( x1 + x2 + x3 == 0
                  , x1 + x2 >= 0
                  , x3 >= 0
                  )
simplify_fixed_values(rules)
}
}
